refactor: move user api to v2
